added scrollable view of input matches
