+++ hw3/03-react/src/components/search/search.css
@@ -9,10 +9,9 @@
 }
 
 .search-input {
-  margin: 1rem auto;
-  width: 13rem;
+  margin: 1rem auto 0 auto;
+  width: 15rem;
   height: 2rem;
-  border-radius: 0.5em;
 }
 
 .card {
@@ -23,6 +22,14 @@
 
 .input-error {
   color: var(--whitetail);
+}
+
+.dropdown-list {
+  margin: 0 auto;
+  max-height: 7rem;
+  width: 15rem;
+  overflow: auto; 
+  background: var(--black_swan);
 }
 
 .name {
